backend/api/requirements.txt
# ============================================
# BACKEND: api/requirements.txt
# ============================================
# Versions matched with training environment (numpy 1.26.4 compatible)

# ===============================
# Core ML/AI (Training-matched)
# ===============================
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
joblib==1.3.2
threadpoolctl==3.2.0

# ===============================
# TensorFlow Lite
# ===============================
# tensorflow==2.13.0
tensorflow>=2.13.0
flatbuffers==23.5.26
protobuf==4.23.4

# ===============================
# Computer Vision
# ===============================
opencv-python==4.8.1.78
Pillow==10.1.0

# ===============================
# Web Framework (FastAPI)
# ===============================
fastapi==0.104.1
uvicorn[standard]==0.24.0

# pydantic==2.5.3

starlette==0.27.0
python-multipart==0.0.6

# ===============================
# Voice Output
# ===============================
pyttsx3==2.99
psutil==5.9.8

# ===============================
# Networking (Auto-discovery)
# ===============================
zeroconf==0.131.0
qrcode==7.4.2

# ===============================
# Utilities
# ===============================
requests==2.31.0
# typing-extensions==4.8.0
click==8.1.7




